guardian: move verified log records to reader
